update files tests to use .component.js convention
