+++ test/files/filelist.component.js
